quick fix query table

+++ BasicBrowser/tmv_app/tasks.py
@@ -190,22 +190,19 @@ and {} topics\n'.format(qid, docs.count(),K))
     del abstracts
     gc.collect()
 
+    if stat.db:
+        vocab = tfidf_vectorizer.get_feature_names()
+        vocab_ids = []
+        pool = Pool(processes=8)
+        vocab_ids.append(pool.map(partial(add_features,run_id=run_id),vocab))
+        pool.terminate()
+        del vocab
+        vocab_ids = vocab_ids[0]
 
-    vocab = tfidf_vectorizer.get_feature_names()
-    vocab_ids = []
-    pool = Pool(processes=8)
-    vocab_ids.append(pool.map(partial(add_features,run_id=run_id),vocab))
-    pool.terminate()
-    del vocab
-    vocab_ids = vocab_ids[0]
-
-
-    ## Make some topics
-    django.db.connections.close_all()
-    topic_ids = db.add_topics(K, run_id)
-
-
-    gc.collect()
+        ## Make some topics
+        django.db.connections.close_all()
+        topic_ids = db.add_topics(K, run_id)
+        gc.collect()
 
     # Fit the NMF model
     print("Fitting the NMF model with tf-idf features, "
@@ -219,76 +216,81 @@ and {} topics\n'.format(qid, docs.count(),K))
     print("done in %0.3fs." % (time() - t0))
     stat.nmf_time = time() - t0
 
-    ## Add topics terms
-    print("Adding topicterms to db")
-    t0 = time()
-    ldalambda = find(csr_matrix(nmf.components_))
-    topics = range(len(ldalambda[0]))
-    tts = []
-    pool = Pool(processes=8)
+    if stat.db:
+        ## Add topics terms
+        print("Adding topicterms to db")
+        t0 = time()
+        ldalambda = find(csr_matrix(nmf.components_))
+        topics = range(len(ldalambda[0]))
+        tts = []
+        pool = Pool(processes=8)
 
-    tts.append(pool.map(partial(db.f_lambda, m=ldalambda,
-                    v_ids=vocab_ids,t_ids=topic_ids,run_id=run_id),topics))
-    pool.terminate()
-    tts = flatten(tts)
-    gc.collect()
-    sys.stdout.flush()
-    django.db.connections.close_all()
-    TopicTerm.objects.bulk_create(tts)
-    print("done in %0.3fs." % (time() - t0))
-    stat.db_time = stat.db_time + time() - t0
-
-
-    ## Add topic-docs
-    gamma =  find(csr_matrix(nmf.transform(tfidf)))
-    glength = len(gamma[0])
-
-    chunk_size = 100000
-
-    ps = 16
-    parallel_add = True
-
-    all_dts = []
-
-    make_t = 0
-    add_t = 0
-
-    t0 = time()
-    ### Go through in chunks
-    for i in range(glength//chunk_size+1):
-        dts = []
-        values_list = []
-        f = i*chunk_size
-        l = (i+1)*chunk_size
-        if l > glength:
-            l = glength
-        docs = range(f,l)
-        doc_batches = []
-        for p in range(ps):
-            doc_batches.append([x for x in docs if x % ps == p])
-        pool = Pool(processes=ps)
-        make_t0 = time()
-        values_list.append(pool.map(partial(
-            db.f_gamma_batch, gamma=gamma,
-            docsizes=docsizes,docUTset=ids,topic_ids=topic_ids,
-            run_id=run_id
-        ),doc_batches))
-        #dts.append(pool.map(partial(f_gamma, gamma=gamma,
-        #                docsizes=docsizes,docUTset=ids,topic_ids=topic_ids),doc_batches))
+        tts.append(pool.map(partial(db.f_lambda, m=ldalambda,
+                        v_ids=vocab_ids,t_ids=topic_ids,run_id=run_id),topics))
         pool.terminate()
-        make_t += time() - make_t0
-        django.db.connections.close_all()
-
-        add_t0 = time()
-        values_list = [item for sublist in values_list for item in sublist]
-        pool = Pool(processes=ps)
-        pool.map(insert_many,values_list)
-        pool.terminate()
-        add_t += time() - add_t0
+        tts = flatten(tts)
         gc.collect()
         sys.stdout.flush()
+        django.db.connections.close_all()
+        TopicTerm.objects.bulk_create(tts)
+        print("done in %0.3fs." % (time() - t0))
+        stat.db_time = stat.db_time + time() - t0
 
-    stat.db_time = stat.db_time + time() - t0
+
+        ## Add topic-docs
+        gamma =  find(csr_matrix(nmf.transform(tfidf)))
+        glength = len(gamma[0])
+
+        chunk_size = 100000
+
+        ps = 16
+        parallel_add = True
+
+        all_dts = []
+
+        make_t = 0
+        add_t = 0
+
+        t0 = time()
+        ### Go through in chunks
+        for i in range(glength//chunk_size+1):
+            dts = []
+            values_list = []
+            f = i*chunk_size
+            l = (i+1)*chunk_size
+            if l > glength:
+                l = glength
+            docs = range(f,l)
+            doc_batches = []
+            for p in range(ps):
+                doc_batches.append([x for x in docs if x % ps == p])
+            pool = Pool(processes=ps)
+            make_t0 = time()
+            values_list.append(pool.map(partial(
+                db.f_gamma_batch, gamma=gamma,
+                docsizes=docsizes,docUTset=ids,topic_ids=topic_ids,
+                run_id=run_id
+            ),doc_batches))
+            #dts.append(pool.map(partial(f_gamma, gamma=gamma,
+            #                docsizes=docsizes,docUTset=ids,topic_ids=topic_ids),doc_batches))
+            pool.terminate()
+            make_t += time() - make_t0
+            django.db.connections.close_all()
+
+            add_t0 = time()
+            values_list = [item for sublist in values_list for item in sublist]
+            pool = Pool(processes=ps)
+            pool.map(insert_many,values_list)
+            pool.terminate()
+            add_t += time() - add_t0
+            gc.collect()
+            sys.stdout.flush()
+
+        stat.db_time = stat.db_time + time() - t0
+
+    em = np.diff(csr_matrix(nmf.transform(tfidf)).indptr) != 0
+
+    stat.empty_topics = len(em[em==False])
 
     stat.error = nmf.reconstruction_err_
     stat.errortype = "Frobenius"
@@ -296,4 +298,5 @@ and {} topics\n'.format(qid, docs.count(),K))
     stat.last_update=timezone.now()
     stat.status=3
     stat.save()
-    management.call_command('update_run',run_id)
+    if stat.db:
+        management.call_command('update_run',run_id)
